Refactor: FieldOffsetStrategy
